+++ modules/console_foundation/outputs.tf
@@ -27,6 +27,8 @@ output "reviewer_role_arn" { value = aws_iam_role.reviewer.arn }
 output "reviewer_role_name" { value = aws_iam_role.reviewer.name }
 output "admin_role_arn" { value = aws_iam_role.admin.arn }
 output "admin_role_name" { value = aws_iam_role.admin.name }
+output "auditor_role_arn" { value = aws_iam_role.auditor.arn }
+output "auditor_role_name" { value = aws_iam_role.auditor.name }
 
 output "console_url" {
   value = "https://${aws_cloudfront_distribution.site.domain_name}"